Magellan: move to using the core for affine transforms
