+++ plugins/Magellan/src/main/java/org/micromanager/magellan/gui/StartupHelpWindow.java
@@ -16,7 +16,7 @@
 //
 package org.micromanager.magellan.gui;
 
-import org.micromanager.magellan.coordinates.AffineUtils;
+import org.micromanager.magellan.coordinates.MagellanAffineUtils;
 import java.awt.Color;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
